chore: feature 폴더명 features로 변경 및 로그인 query 분리
