+++ src/pages/ResetPage.tsx
@@ -1,8 +1,8 @@
-import ResetComplete from '../feature/auth/reset/components/ResetComplete';
-import ResetEmail from '../feature/auth/reset/components/ResetEmail';
-import ResetPassword from '../feature/auth/reset/components/ResetPassword';
-import ResetVerification from '../feature/auth/reset/components/ResetVerification';
-import { usePatchResetPassword } from '../feature/auth/reset/remotes/query';
+import ResetComplete from '../features/auth/reset/components/ResetComplete';
+import ResetEmail from '../features/auth/reset/components/ResetEmail';
+import ResetPassword from '../features/auth/reset/components/ResetPassword';
+import ResetVerification from '../features/auth/reset/components/ResetVerification';
+import { usePatchResetPassword } from '../features/auth/reset/remotes/query';
 import { useFunnel } from '../hooks/use-funnel/useFunnel';
 import { FormProvider, useForm } from 'react-hook-form';
