+++ src/pages/SignupPage.tsx
@@ -1,8 +1,8 @@
-import SignupComplete from '../feature/auth/signup/components/SignupComplete';
-import SignupEmail from '../feature/auth/signup/components/SignupEmail';
-import SignupPassword from '../feature/auth/signup/components/SignupPassword';
-import SignupVerification from '../feature/auth/signup/components/SignupVerification';
-import useSignup from '../feature/auth/signup/remotes/hooks/useSignup';
+import SignupComplete from '../features/auth/signup/components/SignupComplete';
+import SignupEmail from '../features/auth/signup/components/SignupEmail';
+import SignupPassword from '../features/auth/signup/components/SignupPassword';
+import SignupVerification from '../features/auth/signup/components/SignupVerification';
+import useSignup from '../features/auth/signup/remotes/hooks/useSignup';
 import { useFunnel } from '../hooks/use-funnel/useFunnel';
 import { FormProvider, useForm } from 'react-hook-form';
